experimental/requirements_simple.txt
flask==2.3.3
flask-cors==4.0.0
openai==0.28.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
gunicorn==21.2.0
pdfplumber==0.10.0
